Prints all the elements of a list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,5 +1,8 @@
 #include "lists.h"
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
 /**
  * print_listint - function that prints all the elements of a list
  * @h: list
@@ -8,7 +11,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t nodes = 0;
+	int nodes = 0;
 
 	while (h != NULL)
 	{
